modularized BART API
